Implemented CRUD for holiday requests

+++ backend/HR System/EmployeeService.Infrastructure/Persistence/Employee/EmployeeRepository.cs
@@ -56,10 +56,10 @@ namespace EmployeeService.Infrastructure.Persistance.Employee
             return await _context.Employees.AsNoTracking().FirstOrDefaultAsync(x => x.Id == id, cancellationToken);
         }
 
-        public async Task<decimal> GetTotalPagesAsync(int page, int items, CancellationToken cancellationToken = default)
+        public async Task<int> GetTotalPagesAsync(int page, int items, CancellationToken cancellationToken = default)
         {
             var count = await _context.Employees.CountAsync(cancellationToken);
-            return Math.Floor((decimal)((decimal)count / (decimal)items));
+            return (int)Math.Floor((double)count / items);
         }
 
         public async Task<Core.Entities.Employee?> UpdateEmployeeAsync(Core.Entities.Employee user, CancellationToken cancellationToken = default)
